requirements.txt
openpyxl >= 2.3.5
pyYAML >= 3.11
pycurl
